refactor: Move Coefficients and TransformCoefficients to separate classes
