+++ render-app/src/main/java/org/janelia/alignment/filter/LinearIntensityMap8BitFilter.java
@@ -8,7 +8,7 @@ import ij.process.ImageProcessor;
 
 import net.imglib2.algorithm.blocks.BlockSupplier;
 import net.imglib2.type.numeric.integer.UnsignedByteType;
-import org.janelia.alignment.intensity.FastLinearIntensityMap.Coefficients;
+import org.janelia.alignment.intensity.Coefficients;
 import org.janelia.alignment.intensity.LinearIntensityMap;
 
 import net.imglib2.img.Img;
